Update tests which mix int and float types without casts.
These need to change because type coercion in SkSL is about to become
more strict in a followup CL; we are disallowing expressions that mix
ints and floats without a cast.

Bug: skia:11164
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/353102
Commit-Queue: John Stiles <johnstiles@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>

+++ tests/sksl/shared/golden/Ossfuzz28794.metal
@@ -10,7 +10,7 @@ fragment Outputs fragmentMain(Inputs _in [[stage_in]], bool _frontFacing [[front
     Outputs _outputStruct;
     thread Outputs* _out = &_outputStruct;
     int i = int(sqrt(1.0));
-    float(i) * float(i = 3);
+    i * int(int(i = 3));
     _out->sk_FragColor.x = 3.0;
     return *_out;
 }
